added condest to linalg.basics, including tests
